% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/apiDF.R
\name{apiDF}
\alias{apiDF}
\title{Creates data frame after data is called using blsAPI.R}
\usage{
apiDF(data)
}
\arguments{
\item{data}{The JSON used to extract the data gathered from the blsAPI function}
}
\value{
returns a data frame of the data requested from the bLSAPI function call
}
\description{
Used in the laus_get_data function
}
\examples{
library(blsAPI)
library(rjson)
response <- blsAPI('LAUCN040010000000005')
json <- fromJSON(response)
df <- apiDF(json$Results$series[[1]]$data)

}
