% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/blsAPI.R
\name{blsAPI}
\alias{blsAPI}
\title{Request Data From The Bureau Of Labor Statistics' API}
\usage{
blsAPI(payload = NA, api.version = 1)
}
\arguments{
\item{payload}{a string or a list containing data to be sent to the API.}

\item{api.version}{an integer for which api version you want to use (i.e. 1 for v1, 2 for v2)}
}
\description{
Allows users to request data for one or multiple series through the BLS API.  Users provide parameters as specified in \url{http://www.bls.gov/developers/api_signature.htm} and the function returns a JSON string.
}
\details{
See \url{http://www.bls.gov/developers/} and \url{http://www.bls.gov/developers/api_signature.htm} for more details.
}
\examples{
#' ## These examples are taken from http://www.bls.gov/developers/api_signature.htm
library(rjson)

## Single Series request
response <- blsAPI('LAUCN040010000000005')
json <- fromJSON(response)

## Multiple Series
payload <- list('seriesid'=c('LAUCN040010000000005','LAUCN040010000000006'))
response <- blsAPI(payload)
json <- fromJSON(response)

## One or More Series, Specifying Years
payload <- list(
 'seriesid'=c('LAUCN040010000000005','LAUCN040010000000006'),
 'startyear'=2010,
 'endyear'=2012)
response <- blsAPI(payload)
json <- fromJSON(response)
}
\keyword{api}
\keyword{bls}
\keyword{economics}

