% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/help_laus_areacodes.R
\name{help_laus_areacodes}
\alias{help_laus_areacodes}
\title{Prints a list of area names associated to an area code for the LAUS data}
\usage{
help_laus_areacodes()
}
\value{
prints a list in the console
}
\description{
Prints a list of area names associated to an area code for the LAUS data
}
\examples{
library(blsAPI)
library(dplyr)
help_laus_areacodes() 

}
