% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/LAUS_AreaCodes.R
\docType{data}
\name{LAUS_AreaCodes}
\alias{LAUS_AreaCodes}
\title{Area codes for Labor Area Unemployment Statistics (LAUS) data}
\format{
A data frame with 8321 rows and 2 variables
\describe{
    \item{area_code}{area code for LAUS}
    \item{area_text}{area name for LAUS}
}
}
\source{
\url{https://download.bls.gov/pub/time.series/la/la.area}
}
\usage{
LAUS_AreaCodes
}
\description{
A data set containing the area code and area text associated with each code for the LAUS data set
}
\keyword{datasets}
