% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/laus_get_areacode.R
\name{laus_get_areacode}
\alias{laus_get_areacode}
\title{Acquire the area code used in the series id for observations in Labor Area Unemployment Statistics (LAUS) data}
\usage{
laus_get_areacode(Location_Name)
}
\arguments{
\item{Location_Name}{A string or vector of the different cities, states or metropolitan statistical areas you want LAUS data from}
}
\value{
returns a string or vector representing area codes
}
\description{
Gathers the area code used in the laus_get_data function to acquire a data frame using the blsAPI function that request data through the U.S. Bureau of Labor Statistics API. The laus_get_areacode function is called by the laus_get_data function not the user.
}
\details{
type help_laus_areacodes() to get a list of all of the area names that area associated to an area code in the LAUS data
}
\examples{
library(blsAPI)
laus_get_areacode(Location_Name = c("Florida", "California", 
                    "Charlotte County, FL", "Fresno County, CA"))

}
