% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/laus_get_data.R
\name{laus_get_data}
\alias{laus_get_data}
\title{A wrapper function for blsAPI.R function that processes gathered Labor Are Unemployment Statistics (LAUS) data into a data frame}
\usage{
laus_get_data(
  location.vector,
  measure.vector,
  start.year,
  end.year,
  api.version = 1,
  bls.key = NULL
)
}
\arguments{
\item{location.vector}{}

\item{measure.vector}{}

\item{start.year}{}

\item{end.year}{}

\item{api.version}{}

\item{bls.key}{}
}
\description{
Allows users to request LAUS data and have it downloaded as a data frame with ease
}
\examples{
library(blsAPI)
library(dplyr)
library(rjson)
unem_df <- laus_get_data(location.vector = c("Florida", "California", 
                    "Charlotte County, FL", "Fresno County, CA"), 
                     measure.vector = "unemployment rate", 
                     start.year = 2019, end.year = 2021, 
                     api.version = 1)

}
