% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/laus_get_measure.R
\name{laus_get_measure}
\alias{laus_get_measure}
\title{Acquire the measure code used in the series id for observations in Labor Area Unemployment Statistics (LAUS) data}
\usage{
laus_get_measure(measure)
}
\arguments{
\item{measure}{a string containing the desired measure. Exs: unemployment rate, labor force, employment, etc.}
}
\value{
returns a string representing a measure code
}
\description{
Gathers the measure code used in the laus_get_data function to acquire a data frame using the blsAPI function that request data through the U.S. Bureau of Labor Statistics API. The laus_get_areacode function is called by the laus_get_data function not the user.
}
\details{
See <\url{https://download.bls.gov/pub/time.series/la/la.measure}> to see the format of the strings used in the measure param found in the measure_text column
}
\examples{
library(blsAPI)
laus_get_measure("unemployment rate")
}
