% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/laus_measure.R
\docType{data}
\name{laus_measure}
\alias{laus_measure}
\title{Measure codes for Labor Area Unemployment Statistics (LAUS) data}
\format{
A data frame with 7 rows and 2 variables
\describe{
    \item{measure_code}{measure code for LAUS}
    \item{measure_text}{measure name for LAUS}
}
}
\source{
\url{https://download.bls.gov/pub/time.series/la/la.measure}
}
\usage{
laus_measure
}
\description{
A data set containing the measure code and measure text associated with each code for the LAUS data set
}
\keyword{datasets}
