% Generated by roxygen2 (4.0.1): do not edit by hand
\docType{package}
\name{blsAPI}
\alias{blsAPI}
\alias{blsAPI-package}
\title{blsAPI}
\usage{
blsAPI(data = NA)
}
\arguments{
\item{data}{Information to send to the API.}
}
\description{
This function allows users to request data from the BLS API.
See http://www.bls.gov/developers/ and http://www.bls.gov/developers/api_signature.htm for more details.
}
\examples{
## These examples are taken from http://www.bls.gov/developers/api_signature.htm

## Single Series request
response <- blsAPI('LAUCN040010000000005')
json <- fromJSON(response)

## Multiple Series
payload <- list('seriesid'=c('LAUCN040010000000005','LAUCN040010000000006'))
response <- blsAPI(payload)
json <- fromJSON(response)

## One or More Series, Specifying Years
payload <- list('seriesid'=c('LAUCN040010000000005','LAUCN040010000000006'), 'startyear'='2010', 'endyear'='2012')
response <- blsAPI(payload)
json <- fromJSON(response)
}
\keyword{api}
\keyword{bls}
\keyword{economics}

