% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/blsAPI.R
\name{blsAPI}
\alias{blsAPI}
\title{Request data from the Bureau of Labor Statistics' API}
\usage{
blsAPI(payload = NA)
}
\arguments{
\item{payload}{data sent to the API.}
}
\description{
Allows users to request data for one or multiple series through the BLS API.  Users provide parameters as specified in \url{http://www.bls.gov/developers/api_signature.htm} and the function returns a JSON string.
}
\details{
See \url{http://www.bls.gov/developers/} and \url{http://www.bls.gov/developers/api_signature.htm} for more details.
}
\examples{
## These examples are taken from http://www.bls.gov/developers/api_signature.htm

## Single Series request
response <- blsAPI('LAUCN040010000000005')
json <- fromJSON(response)

## Multiple Series
payload <- list('seriesid'=c('LAUCN040010000000005','LAUCN040010000000006'))
response <- blsAPI(payload)
json <- fromJSON(response)

## One or More Series, Specifying Years
payload <- list(
 'seriesid'=c('LAUCN040010000000005','LAUCN040010000000006'),
 'startyear'=2010,
 'endyear'=2012)
response <- blsAPI(payload)
json <- fromJSON(response)
}
\keyword{api}
\keyword{bls}
\keyword{economics}

