% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/laus_get_data.R
\name{laus_get_data}
\alias{laus_get_data}
\title{A wrapper function for blsAPI.R function that processes gathered Labor Are Unemployment Statistics (LAUS) data into a data frame}
\usage{
laus_get_data(
  location.vector,
  measure.vector,
  start.year,
  end.year,
  api.version = 1,
  bls.key = NULL
)
}
\arguments{
\item{location.vector}{A string or vector of the different cities, states or metropolitan statistical areas you want LAUS data from}

\item{measure.vector}{A string of the laus measure you want to gather in your call, e.g. unemployment, unemployment rate.}

\item{start.year}{The year you want as the beginning period of data collection}

\item{end.year}{The year you want as the ending period of data collection}

\item{api.version}{A numerical value that specifies which version of the api you're using (1 or 2). Default is version 1.}

\item{bls.key}{The BLS key you're using to retrieve data using version 2}
}
\description{
Allows users to request LAUS data and have it downloaded as a data frame with ease
}
\examples{
library(blsAPI)
library(dplyr)
library(rjson)
unem_df <- laus_get_data(location.vector = c("Florida", "California", 
                    "Charlotte County, FL", "Fresno County, CA"), 
                     measure.vector = "unemployment rate", 
                     start.year = 2019, end.year = 2021, 
                     api.version = 1)

}
